+++ faddr/__init__.py
@@ -1,20 +1,13 @@
 """Parse network devices' configuration and store in database."""
 
-import sys
 import os
+import sys
 
 from loguru import logger
 
-from faddr.rancid import RancidDir, RancidGroup
-
-
 __version__ = "0.0.4"
-__all__ = (
-    "RancidDir",
-    "RancidGroup",
-)
 
-if os.getenv("FADDR_DEBUG"):
+if os.getenv("FADDR_DEBUG") or any(arg in sys.argv for arg in ("-d", "--debug")):
     LOG_LEVEL = "DEBUG"
 else:
     LOG_LEVEL = "INFO"